thisCode/requirements.txt
absl-py==0.9.0
aiodns==2.0.0
aiohttp==3.6.2
asgiref==3.2.7
astunparse==1.6.3
async-timeout==3.0.1
attrs==19.3.0
bson==0.5.8
cachetools==4.1.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cycler==0.10.0
Cython==0.29.20
dataclasses==0.6
Django==2.0
django-mongodb-engine==0.6.0
djangorestframework==3.11.0
djangotoolbox==1.8.0
django-webpack-loader==0.7.0
djongo==1.2.30
dnspython==1.16.0
gast==0.3.3
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
idna==2.9
imageai==2.1.5
imutils==0.5.3
install==1.3.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-resnet==0.2.0
keras-retinanet==0.5.1
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.2.2
mongoengine==0.20.0
multidict==4.7.6
numpy==1.19.0
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
Pillow==7.1.2
# pip==20.1.1
progressbar2==3.51.4
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==3.1.1
pycparser==2.20
pymongo==3.10.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-utils==2.4.0
pytz==2018.5
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
setuptools==46.1.3
six==1.14.0
slack==0.0.2
slack-client==0.3.0
slackclient==2.0.0
slackclient-cli==0.4.6
sqlparse==0.2.4
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
tensorflow-gpu-estimator==2.2.0
termcolor==1.1.0
urllib3==1.25.9
utils==1.0.1
Werkzeug==1.0.1
wheel==0.34.2
wincertstore==0.2
wrapt==1.12.1
yarl==1.4.2
mod-wsgi==4.7.1
openpyxl==3.0.4